+++ internal/agent/command_test.go
