+++ internal/otel/components.go
