+++ internal/otel/config/config.go
@@ -3,10 +3,10 @@ package config
 import (
 	"errors"
 
-	"github.com/hashicorp/consul-telemetry-collector/pkg/otel/config/helpers/exporters"
-	"github.com/hashicorp/consul-telemetry-collector/pkg/otel/config/helpers/extensions"
-	"github.com/hashicorp/consul-telemetry-collector/pkg/otel/config/helpers/processors"
-	"github.com/hashicorp/consul-telemetry-collector/pkg/otel/config/helpers/receivers"
+	"github.com/hashicorp/consul-telemetry-collector/internal/otel/config/helpers/exporters"
+	"github.com/hashicorp/consul-telemetry-collector/internal/otel/config/helpers/extensions"
+	"github.com/hashicorp/consul-telemetry-collector/internal/otel/config/helpers/processors"
+	"github.com/hashicorp/consul-telemetry-collector/internal/otel/config/helpers/receivers"
 	"github.com/hashicorp/go-multierror"
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/service"
