+++ internal/otel/config/doc.go
