+++ internal/otel/config/helpers/extensions/oauth_client_extension_test.go
