+++ internal/otel/providers/external/external_test.go
