+++ internal/otel/testdata/hcp-with-forwarder.yaml
