+++ internal/otel/testdata/hcp.yaml
