+++ internal/otel/testdata/stock-with-forwarder.yaml
