+++ internal/otel/testdata/stock.yaml
